Create context for teacher dashboard.
