+++ src/components/teacherDashboard/subjectList/index.ts
@@ -1,0 +1,1 @@
+export { SubjectList } from "./SubjectList";
