--- src/components/teacherDashboard/subjectListCard/index.ts
@@ -1,1 +1,0 @@
-export { SubjectListCard } from "./SubjectListCard";
