+++ src/components/teacherDashboard/subjectManager/index.ts
@@ -1,0 +1,1 @@
+export { SubjectManager } from "./SubjectManager";
