+++ src/components/teacherDashboard/topicManager/index.ts
@@ -1,0 +1,1 @@
+export { TopicManager } from "./TopicManager";
